Buildable with aarch64-linux-*-gcc
Modified the build script to deal with gcc that defaults to -fPIC (generating
a global offset table in the elf file).

Reorganised how memset is brought into the compiliation.

Removed an assumption about how the address of _start would be calculated.
This might be better done differently, but the EL3 and EL1 kernels run at
different addresses. PC-relative is still the way to go.

Still doesn't run without a modified qemu that loads the "bios" at 0 and
allocates fb memory on a 2MiB boundary.

Changed the build when generating for use with qemu; afaics it still doesn't
allow for an image file to be treated as an SD card at the hardware level.

+++ el3.c
@@ -3,7 +3,7 @@
 // This file is the entry point for EL3, it performs initialisation, then drops to Secure EL1
 
 // memset, the only C library function used in the kernel (mostly because the optimiser uses it).
-void *memset(void *s, int c, unsigned long long n);
+void *memset(void *s, int c, long unsigned int n);
 
 #include "kernel.h"
 #include "exclusive.h"
@@ -109,7 +109,7 @@ void __attribute__(( noreturn, noinline )) c_el3_nommu( Core *core, unsigned num
   // But use local variables with care, if at all!
   memset( core, 0, sizeof( Core ) );
 
-if (number != 0) { for (;;) { asm ( "wfi" ); } }
+if (number != 0) { for (;;) { asm ( "wfi" ); } } // TODO TODO TODO!!
   // Note: the address of the present array will be an offset from _start.
   el3_synchronised_initialise( core, number, present_bits );
