+++ include/memset.h
@@ -1,1 +1,2 @@
-void *memset(void *s, int c, unsigned long long n);
+
+void *memset(void *s, int c, unsigned long int n);
