+++ memset.c
@@ -1,9 +1,8 @@
-#include <stddef.h>
 
-void *memset(void *s, int c, size_t n)
+void *memset(void *s, int c, long unsigned int n)
 {
   char *p = s;
-  for (size_t i = 0; i < n; i++) {
+  for (unsigned long int i = 0; i < n; i++) {
     p[i] = c;
   }
   return s;
